server/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
